Extract graph modal functions and forms to separate files
