+++ app/frontend/src/graph.js
@@ -1,14 +1,15 @@
 import van from "vanjs-core"
-import {Modal} from "vanjs-ui"
 import { v4 as uuidv4 } from 'uuid';
 import hash from 'object-hash';
 
-import {Get, Put} from './backend.js';
-import {DoAppErrorModal} from './apperror.js';
-import {ButtonAct, ButtonCancel} from './buttons.js';
-import {Table, TableRow} from './table.js';
+import { Get, Put } from './backend.js';
+import { DoAppErrorModal} from './apperror.js';
+import { DoBlockModal } from "./block.js";
+import { ButtonAct } from './buttons.js';
+import { DoConnectionModal } from "./connection.js";
+import { Table, TableRow } from './table.js';
 
-const {div, input, label, p, tbody} = van.tags
+const {div, p, tbody} = van.tags
 
 const getGraph = async (id) => {
     console.log(`getting graph ${id}`);
@@ -95,140 +96,6 @@ const ConnTableRow = ({source, target, onView, onDelete}) => {
     const rowItems = [source, target, buttons];
 
     return () => deleted.val ? null : TableRow(rowItems);
-}
-
-const BlockForm = ({name, type, onOK, onCancel}) => {
-    const inputClass = "block px-5 py-5 mt-2 border border-gray-200 rounded-md focus:border-indigo-500 focus:outline-none focus:ring";
-
-    return div(
-        {class: "flex flex-col drop-shadow hover:drop-shadow-lg w-300 rounded-md"},
-        p({class: "text-lg font-medium font-bold text-center"}, "Graph Block"),
-        div(
-            div(
-                label({for: "name"}, "Name"),
-                input({
-                    id: "name",
-                    class: inputClass,
-                    type: "text",
-                    value: name,
-                    placeholder: "Non-empty, unique",
-                    oninput: e => name.val = e.target.value,
-                }),
-                label({for: "type"}, "Type"),
-                input({
-                    id: "type",
-                    class: inputClass,
-                    type: "text",
-                    value: type,
-                    placeholder: "Valid block type",
-                    oninput: e => type.val = e.target.value,
-                }),
-            ),
-        ),
-        div(
-            {class:"mt-4 flex justify-center"},
-            ButtonCancel({text: "Cancel", onclick: onCancel}),
-            ButtonAct({text: "OK", onclick: onOK}),
-        ),
-    )
-}
-
-const ConnectionForm = ({source, target, onOK, onCancel}) => {
-    const inputClass = "block px-5 py-5 mt-2 border border-gray-200 rounded-md focus:border-indigo-500 focus:outline-none focus:ring";
-
-    return div(
-        {class: "flex flex-col drop-shadow hover:drop-shadow-lg w-300 rounded-md"},
-        p({class: "text-lg font-medium font-bold text-center"}, "Graph Connection"),
-        div(
-            div(
-                label({for: "source"}, "Source Address"),
-                input({
-                    id: "source",
-                    class: inputClass,
-                    type: "text",
-                    value: source,
-                    placeholder: "<block.output>",
-                    oninput: e => source.val = e.target.value,
-                }),
-                label({for: "target"}, "Target Address"),
-                input({
-                    id: "target",
-                    class: inputClass,
-                    type: "text",
-                    value: target,
-                    placeholder: "<block.input>",
-                    oninput: e => target.val = e.target.value,
-                }),
-            ),
-        ),
-        div(
-            {class:"mt-4 flex justify-center"},
-            ButtonCancel({text: "Cancel", onclick: onCancel}),
-            ButtonAct({text: "OK", onclick: onOK}),
-        ),
-    )
-}
-
-const DoBlockModal = ({block, handleResult}) => {
-    const closed = van.state(false)
-
-    const name = van.state(block.name);
-    const type = van.state(block.type);
-    const paramVals = van.state(block.paramVals);
-    const recording = van.state(block.recording);
-
-    van.add(
-        document.body,
-        Modal({closed},
-            BlockForm({
-                name: name,
-                type: type,
-                // paramVals: paramVals,
-                // recording: recording,
-                onOK: ()=> {
-                    handleResult({
-                        name: name.val,
-                        type: type.val,
-                        paramVals: paramVals.val,
-                        recording: recording.val,
-                    });
-
-                    closed.val = true;
-                },
-                onCancel: () => {
-                    closed.val = true;
-                }
-            }),
-        ),
-    );
-}
-
-const DoConnectionModal = ({connection, handleResult}) => {
-    const closed = van.state(false);
-    const source = van.state(connection.source);
-    const target = van.state(connection.target);
-
-    van.add(
-        document.body,
-        Modal({closed},
-            ConnectionForm({
-                source: source,
-                target: target,
-                onOK: ()=> {
-                    console.log("pressed OK")
-
-                    handleResult({source: source.val, target: target.val});
-
-                    console.log("closing modal")
-
-                    closed.val = true;
-                },
-                onCancel: () => {
-                    closed.val = true;
-                }
-            }),
-        ),
-    );
 }
 
 const Graph = (id) => {
